requirements.txt
PyYAML
numpy===1.19.3
transformers
sentence-transformers
